Set up application
Currently this isn't set up to receive and handle the results from Zillow.
Their API seems hell-bent on rejecting incoming requests and sending them
to a CAPTCHA page.  This could have something to do with how I was testing it
or something else, and there's no real way of knowing.